split correctly the presentations in three seperate folders
