+++ examples/app/interactivePresentation/views/Question1View.scala
@@ -1,6 +1,6 @@
-package views.interactivePresentation
+package interactivePresentation.views
 
-object SubmissionConfirmationView {
+object Question1View {
 
   import dsl.Dsl._
 
@@ -9,7 +9,7 @@ object SubmissionConfirmationView {
   def apply(title: String, description: String, author: String, theme: String) = {
     presentation(title, description, author, theme,
       slide(
-        title3("Thank you for your answer")
+        surveyOf2("Question1", "Are You Paying Attention ?","Yes", "No")
       )
     )
   }
