+++ examples/conf/routes
@@ -4,19 +4,19 @@
 
 
 # Home page
-GET     /                                           @controllers.PresentationController.index
+GET     /                                           @reactiveSlides.controllers.PresentationController.index
 
 # The Reveal.js demo Presentation routes
-GET     /RevealJsDemo                               @controllers.PresentationController.displayRevealJsDemo
+GET     /RevealJsDemo                               @revealJsDemoPresentation.controllers.revealJsPresentationController.displayRevealJsDemo
 
 # The Interactive Presentation routes
-GET     /InteractiveDemo                            @controllers.PresentationController.displayInteractiveDemo
-GET     /Question1                                  @controllers.PresentationController.showQuestion1
-GET     /Question2                                  @controllers.PresentationController.showQuestion2
-POST    /SubmitAnswer                               @controllers.PresentationController.saveAnswer
+GET     /InteractiveDemo                            @interactivePresentation.controllers.interactivePresentationController.displayInteractiveDemo
+GET     /Question1                                  @interactivePresentation.controllers.interactivePresentationController.showQuestion1
+GET     /Question2                                  @interactivePresentation.controllers.interactivePresentationController.showQuestion2
+POST    /SubmitAnswer                               @interactivePresentation.controllers.interactivePresentationController.saveAnswer
 
 # The example Presentation routes
-GET     /ExampleDemo                                @controllers.PresentationController.displayExampleDemo
+GET     /ExampleDemo                                @examplePresentation.controllers.examplePresentationController.displayExampleDemo
 
 # Map static resources from the /public folder to the /assets URL path
 GET     /assets/*file               controllers.Assets.at(path="/public", file)
